third_party/blink/web_tests/platform/generic/storage/indexeddb/mozilla/key-requirements-inline-and-passed-expected.txt

Test IndexedDB's behavior adding inline and passed key simultaneously

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

dbname = "key-requirements-inline-and-passed.html"
indexedDB.deleteDatabase(dbname)
indexedDB.open(dbname)
objectStore = db.createObjectStore('baz', { keyPath: 'id' });
Expecting exception from objectStore.add({id: 5}, 5);
PASS Exception was thrown.
PASS code is 0
PASS ename is 'DataError'
Exception message: Failed to execute 'add' on 'IDBObjectStore': The object store uses in-line keys and the key parameter was provided.
PASS successfullyParsed is true

TEST COMPLETE
